test: move fake fesom mesh to fixture
