+++ conftest.py
@@ -7,15 +7,16 @@ import pytest
 from tests.utils.constants import TEST_ROOT
 
 pytest_plugins = [
+    "tests.fixtures.CMIP_Tables_Dir",
+    "tests.fixtures.CV_Dir",
+    "tests.fixtures.config_files",
     "tests.fixtures.configs",
     "tests.fixtures.environment",
-    "tests.fixtures.fake_filesystem",
-    "tests.fixtures.sample_rules",
-    "tests.fixtures.config_files",
-    "tests.fixtures.CV_Dir",
-    "tests.fixtures.CMIP_Tables_Dir",
     "tests.fixtures.example_data.fesom_2p6_pimesh",
     "tests.fixtures.example_data.pi_uxarray",
+    "tests.fixtures.fake_data.fesom_mesh",
+    "tests.fixtures.fake_filesystem",
+    "tests.fixtures.sample_rules",
 ]
